+++ packages/falcon-dev-tools/falcon-scripts/src/build-esm.js
@@ -1,0 +1,16 @@
+const spawn = require('cross-spawn');
+const Logger = require('@deity/falcon-logger');
+
+module.exports = () => {
+  Logger.log('building esm...');
+
+  const babelConfigPath = require.resolve('./babel/babel.config');
+
+  const result = spawn.sync(`babel`, ['src', '-d', 'dist', '-x', '.ts,.tsx', '-s', '--config-file', babelConfigPath], {
+    stdio: 'inherit'
+  });
+
+  if (result.status !== 0) {
+    throw result;
+  }
+};