+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) are already released and available via npm. Versions without a date are not released yet.
 
+## Falcon vNEXT
+
+### Falcon Scripts v0.0.1
+
+- initial release
+
+---
+
 ## Falcon v1.0
 
 ### Falcon Magento2 API v0.5.0 (2019-04-11)